requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
pdfplumber==0.11.0
openai==1.12.0
python-dotenv==1.0.1
pytest==8.0.2
python-multipart==0.0.9
httpx==0.27.0
aiofiles==23.2.1
sqlalchemy==2.0.28
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
email-validator==2.1.1